- Added MovieImage and ImagesResponse data classes, reorganizing packages
