+++ app/src/main/java/com/arctouch/codechallenge/core/api/model/GenresResponse.kt
@@ -1,5 +1,5 @@
 package com.arctouch.codechallenge.core.api.model
 
-import com.arctouch.codechallenge.home.domain.model.Genre
+import com.arctouch.codechallenge.core.domain.model.Genre
 
 data class GenresResponse(val genres: List<Genre>)
