+++ app/src/main/java/com/arctouch/codechallenge/core/api/model/ImagesResponse.kt
@@ -1,0 +1,8 @@
+package com.arctouch.codechallenge.core.api.model
+
+import com.arctouch.codechallenge.core.domain.model.Genre
+import com.arctouch.codechallenge.core.domain.model.MovieImage
+
+data class ImagesResponse(
+        val backdrops: List<Genre>,
+        val posters: List<MovieImage>)
