+++ app/src/main/java/com/arctouch/codechallenge/core/domain/model/Movie.kt
@@ -1,4 +1,4 @@
-package com.arctouch.codechallenge.home.domain.model
+package com.arctouch.codechallenge.core.domain.model
 
 import android.os.Parcelable
 import com.squareup.moshi.Json
@@ -10,7 +10,7 @@ data class Movie(
         val title: String,
         val overview: String?,
         var genres: List<Genre>?,
-        var images: List<String>?,
+        var images: List<MovieImage>?,
         @field:Json(name = "genre_ids") val genreIds: List<Int>?,
         @field:Json(name = "poster_path") val posterPath: String?,
         @field:Json(name = "backdrop_path") val backdropPath: String?,
