+++ app/src/main/java/com/arctouch/codechallenge/core/extensions/MovieExtensions.kt
@@ -1,7 +1,7 @@
 package com.arctouch.codechallenge.core.extensions
 
 import com.arctouch.codechallenge.core.api.TmdbApi
-import com.arctouch.codechallenge.home.domain.model.Movie
+import com.arctouch.codechallenge.core.domain.model.Movie
 
 private const val POSTER_URL = "https://image.tmdb.org/t/p/w500"
 private const val BACKDROP_URL = "https://image.tmdb.org/t/p/w780"
