+++ app/src/main/java/com/arctouch/codechallenge/home/data/repository/MoviesRepository.kt
@@ -1,8 +1,8 @@
 package com.arctouch.codechallenge.home.data.repository
 
 import com.arctouch.codechallenge.core.api.TmdbApi
+import com.arctouch.codechallenge.core.domain.model.Movie
 import com.arctouch.codechallenge.home.data.cache.MoviesCache
-import com.arctouch.codechallenge.home.domain.model.Movie
 
 class MoviesRepository(private val api: TmdbApi) {
