+++ app/src/main/java/com/arctouch/codechallenge/home/presentation/HomeViewModel.kt
@@ -3,7 +3,7 @@ package com.arctouch.codechallenge.home.presentation
 import android.arch.lifecycle.LiveData
 import android.arch.lifecycle.MutableLiveData
 import com.arctouch.codechallenge.core.BaseViewModel
-import com.arctouch.codechallenge.home.domain.model.Movie
+import com.arctouch.codechallenge.core.domain.model.Movie
 import com.arctouch.codechallenge.home.domain.usecase.HomeMoviesUseCase
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.IO
