+++ app/src/main/java/com/arctouch/codechallenge/home/ui/adapter/HomeAdapter.kt
@@ -2,7 +2,7 @@ package com.arctouch.codechallenge.home.ui.adapter
 
 import android.support.v7.widget.RecyclerView
 import android.view.ViewGroup
-import com.arctouch.codechallenge.home.domain.model.Movie
+import com.arctouch.codechallenge.core.domain.model.Movie
 import org.jetbrains.anko.sdk27.listeners.onClick
 
 class HomeAdapter(
